requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
cmdstanpy==0.9.5
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Cython==3.0.0
Django==4.1
djangorestframework==3.14.0
ephem==4.1.4
fonttools==4.42.0
frozendict==2.3.8
holidays==0.10.3
html5lib==1.1
idna==3.4
importlib-resources==6.0.1
joblib==1.3.2
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==4.9.3
matplotlib==3.7.2
multitasking==0.0.11
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
PyMeeus==0.5.12
pyparsing==3.0.9
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
setuptools-git==1.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
yfinance==0.2.28
zipp==3.16.2
